requirements.txt
langchain==0.3.3
MCQGenerator==0.0.1
openai==1.52.0
packaging==24.1
pandas==2.2.3
pillow==10.4.0
protobuf==5.29.0rc1
pyarrow==17.0.0
pydeck==0.9.1
pypdf2==3.0.1
python-dotenv==1.0.1
requests==2.32.3
rich==13.9.2
streamlit==1.39.0
tenacity==9.0.0
toml==0.10.2
tornado==6.4.1
typing-extensions==4.12.2
